requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
elastic-transport==8.17.1
elasticsearch==9.0.2
executing==2.2.0
fastapi==0.115.13
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.1
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
nest-asyncio==1.6.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.2.1
pipx==1.7.1
platformdirs==4.3.8
portalocker==2.10.1
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.3
redis==6.2.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rq==2.4.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
thinc==8.3.6
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
zstandard==0.23.0
